refactor(ui): separate interaction request forms
